+++ src/ui/components/index.ts
@@ -1,7 +1,7 @@
 export { ProgressBar } from "./ProgressBar";
 export { StatusLine } from "./StatusLine";
 export { HeaderLine, type HeaderPrefix } from "./HeaderLine";
-export { InputArea } from "./InputArea";
+export { InputArea, type CommandCandidate } from "./InputArea";
 export { ChatView } from "./ChatView";
 export { MessageBlock } from "./MessageBlock";
 export { ThinkingBlock } from "./ThinkingBlock";